Properly terminate command string
Fix issue when receiving a command string shorter than the previous one.
As there is no NUL-byte (`\0`) sent via the TCP socket, the resulting
command is a concatenation of the new (shorter) string and what's left
of the initial longer one.

Check the number of bytes read and properly place a NUL-byte in the
command string.

Also take care of negative return value from `read()`. And take care of
a possible trailing newline (overwrite that with a NUL-byte).

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>